% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_early_bioc}
\alias{download_early_bioc}
\title{download packages for a version 1.5-1.7}
\usage{
download_early_bioc(vers = "1.5", type = "src")
}
\arguments{
\item{vers}{character(1) defaults to "1.5"}

\item{type}{character(1) in c("src", "win")}
}
\description{
download packages for a version 1.5-1.7
}
