% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_early_bioc}
\alias{scrape_early_bioc}
\title{use rvest to scrape name and version from packages/bioc for early releases}
\usage{
scrape_early_bioc(vers = "1.5", type = "src")
}
\arguments{
\item{vers}{character(1) version "x.y"}
}
\description{
use rvest to scrape name and version from packages/bioc for early releases
}
